# api.yaml
swagger: '2.0'
info:
  description: REST API description of the ElasTest Platform Manager Module.
  version: 0.1.2
  title: EPM REST API
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:8180'
basePath: /v1
paths:
...
  /vdu:
    get:
      tags:
        - VDU
      summary: Returns all VDUs.
      description: Returns all VDUs with all its details.
      operationId: getAllVdus
      produces:
        - application/json
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/VDU'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - VDU
      summary: Allocates resources in the target cloud.
      description: 'Allocates resources defined as a VDU in the cloud to be deployed in the target cloud environment. It instantiates computing resources, deploys artifacts on them and manages the their lifecycle'
      operationId: deployVdu
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Defintion of a VDU which defines resources that have to be deployed
          required: true
          schema:
            $ref: '#/definitions/VDU'
      responses:
        '200':
          description: VDU allocation OK
          schema:
            $ref: '#/definitions/VDU'
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
          schema:
            type: string
  '/vdu/{id}':
    get:
      tags:
        - VDU
      summary: Returns a VDU.
      description: Returns the VDU with the given ID. Returns all its details.
      operationId: getVduById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of VDU
          required: true
          type: string
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/VDU'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: VDU not found.
          schema:
            type: string
    delete:
      tags:
        - VDU
      summary: Terminates a VDU.
      description: Terminates the VDU that matches with a given ID.
      operationId: deleteVdu
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: ID of VDU
          required: true
          type: string
      responses:
        '200':
          description: VDU terminated successfully
          schema:
            type: string
        '204':
          description: No Content
        '400':
          description: Bad Request
          schema:
            type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: VDU not found
          schema:
            type: string
    patch:
      tags:
        - VDU
      summary: Updates a VDU.
      description: Updates an already deployed VDU.
      operationId: updateVdu
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of VDU
          required: true
          type: string
        - in: body
          name: body
          description: VDU object that needs to be updated.
          required: true
          schema:
            $ref: '#/definitions/VDU'
      responses:
        '200':
          description: VDU Update OK
          schema:
            $ref: '#/definitions/VDU'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: VDU not found
...
definitions:
  Event:
    type: object
    required:
      - description
      - timestamp
    properties:
      description:
        type: string
        example: testEvent1
      id:
        type: string
        example: 1234-abcd
      timestamp:
        type: string
        format: date-time
    description: An event contains certain life cycle information of the VDU at a specific time.
  KeyValuePair:
    type: object
    required:
      - key
      - value
    properties:
      id:
        type: string
        example: 1234-abcd
      key:
        type: string
        example: testKey1
      value:
        type: string
        example: testValue1
    description: This entity is a Key-Value pair for storing metadata contained in other entities
...
  VDU:
    description: 'A Virtual Deployment Unit (VDU) describes the capabilities of virtualized computing (Containers, VMs) and networking resources.'
    type: object
    required:
      - computeId
      - imageName
      - ip
      - name
      - netName
      - poPName
    properties:
      computeId:
        type: string
        example: 1234-abcd
      events:
        type: array
        items:
          $ref: '#/definitions/Event'
      id:
        type: string
        example: 1234-abcd
      imageName:
        type: string
        example: testImage1
      ip:
        type: string
        example: 172.0.0.1
      metadata:
        type: array
        items:
          $ref: '#/definitions/KeyValuePair'
      name:
        type: string
        example: testVdu1
      netName:
        type: string
        example: testNetworkName
      poPName:
        type: string
      status:
        type: string
        enum:
          - initializing
          - initialized
          - deploying
          - deployed
          - undeploying
          - undeployed
          - error
...